backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
pillow==10.1.0
razorpay==1.4.1
reportlab
python-dotenv==1.0.0
django-extensions==3.2.3
